GUI Beginning
Created a buttons class that can handle creation of buttons + button methods. Created Dashboard Page which can act as the first page. Start page is currently there as a placeholder to keep functionalitystructure in app.py. will need to rename it and will probably use it as the frame for when we create a first run