More navbars
Renamed navbar.py to side_navbar.py.
Added top_navbar.py which is a frame for the top navbar. Added tkinter listboxes to the side navbar to create a smoother UI experience and will have to link onclick functions to them to switch between views. Postponed working on resize issues in order to work towards features that will be more compelling in the demo.